full stack done

+++ src/main/java/com/gs_data_collector/rest/webservices/restfulwebservices/todo/TodoHardcodedService.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 public class TodoHardcodedService {
 
 	private static List<Todo> todos = new ArrayList();
-	private static int idCounter = 0;
+	private static long idCounter = 0;
 	
 	static {
 		todos.add(new Todo(++idCounter, "adam", "Finish college", new Date(), false));
@@ -34,7 +34,7 @@ public class TodoHardcodedService {
 		return todo;
 	}
 	
-	public Todo deleteById(long id) {
+	public Todo deleteById(Long id) {
 		Todo todo = findById(id);
 		if(todo==null) {
 			return null;
@@ -45,7 +45,7 @@ public class TodoHardcodedService {
 		return null;
 	}
 
-	public Todo findById(long id) {
+	public Todo findById(Long id) {
 		for(Todo todo:todos) {
 			if(todo.getId() == id) {
 				return todo;
